requirements.txt
anyio==4.0.0
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.1.3
flask==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
psycopg2-binary==2.9.9
python-telegram-bot==20.6
requests==2.31.0
requests-oauthlib==0.5.0
sniffio==1.3.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
